Improve quaternion approximations and conversions

+++ CGJ_Project/src/engine/math/quaternion.h
@@ -11,16 +11,16 @@ class Quaternion
 {
 
 public:
-	float _x = 0.0;
-	float _y = 0.0;
-	float _z = 0.0;
-	float _w = 1.0;
+	double _x = 0.0;
+	double _y = 0.0;
+	double _z = 0.0;
+	double _w = 1.0;
 
 
 public:
 	Quaternion() = default;
 	Quaternion(const Quaternion&) = default;
-	Quaternion(float x, float y, float z, float w);
+	Quaternion(double x, double y, double z, double w);
 	Quaternion(const Coords4f& coords);
 	Quaternion(const Coords3f& axis, float angle);
 	Quaternion(const Coords3f& rotation);
@@ -29,8 +29,8 @@ public:
 
 public:
 	inline float length() const							{ return std::sqrt(length2()); }
-	inline float length2() const						{ return _x * _x + _y * _y + _z * _z + _w * _w; }
-	inline float dot(const Quaternion& other) const		{ return _x * other._x + _y * other._y + _z * other._z + _w * other._w; }
+	inline float length2() const						{ return (float)(_x * _x + _y * _y + _z * _z + _w * _w); }
+	inline float dot(const Quaternion& other) const		{ return (float)(_x * other._x + _y * other._y + _z * other._z + _w * other._w); }
 	
 
 public:
